HospitalWebAngular: implementar animación para mensajes de form, botones cargando y disabled y arreglos de button lock flags.

+++ HospitalWebAngular/src/app/modules/administration/pages/doctors/doctors.page.ts
@@ -5,10 +5,9 @@ import { DoctorService } from '@services/doctor/doctor.service';
 import { Field } from '@services/field/field.model';
 import { FieldService } from '@services/field/field.service';
 import { MessageType } from '@services/http/http.types';
-import { ButtonType, InputType, Option } from '@shared/components/form-field/form-field.types';
+import { ButtonType, InputType } from '@shared/components/form-field/form-field.types';
 import { ModalPosition, ModalSize } from '@shared/components/modal/modal.types';
-import { firstValueFrom, map, Observable, of, startWith, tap } from 'rxjs';
-import { runInThisContext } from 'vm';
+import { firstValueFrom, last, Observable, of, startWith, tap } from 'rxjs';
 
 @Component({
   selector: 'app-doctors',
@@ -22,6 +21,7 @@ export class DoctorsPage implements OnInit{
   public modalOpen: boolean;
   public doctor: Doctor;
   public loading: boolean;
+  public saving: boolean;
   public filter: any;
   public confirmText: string;
   public confirmOpen: boolean;
@@ -33,7 +33,6 @@ export class DoctorsPage implements OnInit{
   public ButtonType = ButtonType;
 
   private confirmFunction: () => void;
-  private firstField: Field;
   
   constructor(
     private doctorService: DoctorService,
@@ -45,10 +44,10 @@ export class DoctorsPage implements OnInit{
     this.modalOpen = false;
     this.doctor = null;
     this.loading = false;
+    this.saving = false;
     this.confirmText = null;
     this.confirmOpen = false;
     this.messages = [];
-    this.firstField = null;
     this.confirmFunction = null;
     this.filter = {
       documentId: null,
@@ -67,13 +66,15 @@ export class DoctorsPage implements OnInit{
   }
 
   public list(): void {
-    this.loading = true;
-    this.$doctors = this.doctorService.list(this.filter)
-      .pipe(
-        tap(() => {
-          this.loading = false;
-        })
-      );
+    if(!this.loading) {
+      this.loading = true;
+      this.$doctors = this.doctorService.list(this.filter)
+        .pipe(
+          tap(() => {
+            this.loading = false;
+          })
+        );
+    }
   }
 
   public createUpdate(data: any = null) {
@@ -85,52 +86,58 @@ export class DoctorsPage implements OnInit{
     this.confirmOpen = true;
     this.confirmText = 'Está seguro que desea eliminar este récord?';
     this.confirmFunction = async () => {
-      this.loading = true;
-      const response = await firstValueFrom(this.doctorService.delete(id));
-      if(response != null) {
-        if(response.success) {
-          this.appService.siteMessage = { type: MessageType.Success, text: 'Se eliminó el récord correctamente' };
-          this.list();
-        }else {
-          this.appService.siteMessage = { type: MessageType.Warning, text: response.messages[0] };
+      if(!this.saving) {
+        this.saving = true;
+        const response = await firstValueFrom(this.doctorService.delete(id));
+        if(response != null) {
+          if(response.success) {
+            this.appService.siteMessage = { type: MessageType.Success, text: 'Se eliminó el récord correctamente' };
+            this.list();
+          }else {
+            this.appService.siteMessage = { type: MessageType.Warning, text: response.messages[0] };
+          }
         }
+        this.saving = false;
+        this.confirmOpen = false;
       }
-      this.loading = false;
     }
   }
 
   public async save() {
-    const isNew = this.doctor.id == null
-    const response = await firstValueFrom(isNew ? this.doctorService.post(this.doctor) : this.doctorService.put(this.doctor.id, this.doctor));
-    this.messages = [];
-    if(response != null) {
-      if(response.success) {
-        if(isNew) {
-          this.filter = {
-            documentId: response.data.documentId,
-            firstName: null,
-            lastName: null,
-            fieldId: null 
+    if(!this.saving) {
+      this.saving = true;
+      this.messages = [];
+      
+      const isNew = this.doctor.id == null
+      const response = await firstValueFrom(isNew ? this.doctorService.post(this.doctor) : this.doctorService.put(this.doctor.id, this.doctor));  
+      if(response != null) {
+        if(response.success) {
+          if(isNew) {
+            this.filter = {
+              documentId: response.data.documentId,
+              firstName: null,
+              lastName: null,
+              fieldId: null 
+            }
           }
-        }
 
-        this.modalOpen = false;
-        this.appService.siteMessage = { type: MessageType.Success, text: 'Se guardó el récord correctamente' };
-        this.list();
-      }else {
-        this.messages = response.messages;
+          this.modalOpen = false;
+          this.appService.siteMessage = { type: MessageType.Success, text: 'Se guardó el récord correctamente' };
+          this.list();
+        }else {
+          this.messages = response.messages;
+        }
       }
+      this.saving = false;
     }
-    this.loading = false;
   }
 
   public confirm(confirmed: boolean) {
     if(confirmed) {
       this.confirmFunction();
+    }else {
+      this.confirmOpen = false;
     }
-
-    this.confirmOpen = false;
-
   }
 
   public onConfirmClose() {
